+++ CHANGES.md
@@ -2,6 +2,8 @@
 * GitHub:  <http://github.com/LeventErkok/sbv>
 
 ### Version 11.3.5, Not yet released
+
+  * More exports from Data.SBV.Internals to enable compilation of SBVPlugin.
 ### Version 11.3, 2025-03-10
 
   * Fix various haddock documentation links